+++ src/RouterApp.jsx
@@ -1,22 +1,66 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom'
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 
-import Layout from './layout/Layout'
-import Home from './layout/Home'
-import News from './layout/News'
-import Events from './layout/Events'
-import ContactSection from './layout/ContactSection'
+import Layout from "./blocks/Layout";
+import Home from "./blocks/Home";
+import News from "./blocks/News";
+import Events from "./blocks/Events";
+import ContactSection from "./blocks/ContactBlock";
+import NotFound from "./blocks/NotFound";
 
 function RouterApp() {
   return (
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Layout><Home/></Layout>} />
-          <Route path="/home" element={<Layout><Home/></Layout>} />
-          <Route path="/news" element={<Layout><News/></Layout>} />
-          <Route path="/events" element={<Layout><Events/></Layout>} />
-          <Route path="/contact" element={<Layout><ContactSection/></Layout>} />
-        </Routes>
-      </BrowserRouter>
+    <BrowserRouter>
+      <Routes>
+        <Route
+          path="/"
+          element={
+            <Layout>
+              <Home />
+            </Layout>
+          }
+        />
+        <Route
+          path="/home"
+          element={
+            <Layout>
+              <Home />
+            </Layout>
+          }
+        />
+        <Route
+          path="/news"
+          element={
+            <Layout>
+              <News />
+            </Layout>
+          }
+        />
+        <Route
+          path="/events"
+          element={
+            <Layout>
+              <Events />
+            </Layout>
+          }
+        />
+        <Route
+          path="/contact"
+          element={
+            <Layout>
+              <ContactSection />
+            </Layout>
+          }
+        />
+        <Route
+          path="*"
+          element={
+            <Layout>
+              <NotFound />
+            </Layout>
+          }
+        />
+      </Routes>
+    </BrowserRouter>
   );
 }
 export default RouterApp;
