+++ src/blocks/ContactBlock.jsx
@@ -1,0 +1,12 @@
+import Contact from "../sections/Contact";
+import CustomerLogos from "../sections/CustomerLogos";
+
+const ContactSection = () => {
+  return (
+    <>
+      <Contact />
+      <CustomerLogos />
+    </>
+  );
+};
+export default ContactSection;
