+++ src/blocks/Events.jsx
@@ -1,0 +1,12 @@
+import CtaTwoDonate from "../sections/CtaTwoDonate";
+import Blog from "../sections/Blog";
+
+const Events = () => {
+  return (
+    <>
+      <CtaTwoDonate />
+      <Blog />
+    </>
+  );
+};
+export default Events;
