+++ src/blocks/Home.jsx
@@ -1,0 +1,14 @@
+import Hero from "../sections/Hero";
+import CtaTwo from "../sections/CtaTwo";
+import SocialProof from "../sections/SocialProof";
+
+const Home = () => {
+  return (
+    <>
+      <Hero />
+      <SocialProof />
+      <CtaTwo />
+    </>
+  );
+};
+export default Home;
