+++ src/blocks/Layout.jsx
@@ -1,0 +1,13 @@
+import Navbar from "../sections/Navbar";
+import Footer from "../sections/Footer";
+
+const Layout = ({ children }) => {
+  return (
+    <>
+      <Navbar />
+      {children}
+      <Footer />
+    </>
+  );
+};
+export default Layout;
