+++ src/blocks/News.jsx
@@ -1,0 +1,10 @@
+import CtaOne from "../sections/CtaOne";
+
+const News = () => {
+  return (
+    <>
+      <CtaOne />
+    </>
+  );
+};
+export default News;
