--- src/layout/ContactSection.jsx
@@ -1,12 +1,0 @@
-import Contact from '../contact/Contact'
-import CustomerLogos from '../customerLogos/CustomerLogos'
-
-const ContactSection = () => {
-	return (
-		<>
-			<Contact />
-			<CustomerLogos />
-		</>
-	)
-}
-export default ContactSection;
