--- src/layout/Events.jsx
@@ -1,12 +1,0 @@
-import CtaTwoDonate from '../ctaTwo/CtaTwoDonate'
-import Blog from '../blog/Blog'
-
-const Events = () => {
-	return (
-		<>
-			<CtaTwoDonate />
-			<Blog />
-		</>
-	)
-}
-export default Events;
