--- src/layout/Home.jsx
@@ -1,14 +1,0 @@
-import Hero from '../hero/Hero'
-import CtaTwo from '../ctaTwo/CtaTwo'
-import SocialProof from '../socialProof/SocialProof'
-
-const Home = () => {
-	return (
-		<>
-			<Hero />
-			<SocialProof />
-			<CtaTwo />
-		</>
-	)
-}
-export default Home;
