--- src/layout/Layout.jsx
@@ -1,13 +1,0 @@
-import Navbar from "../navbar/Navbar"
-import Footer from "../footer/Footer"
-
-const Layout = ( {children} ) => {
-	return (
-		<>
-			<Navbar />
-			{children}
-			<Footer />
-		</>
-	);
-}
-export default Layout;
