--- src/layout/News.jsx
@@ -1,10 +1,0 @@
-import CtaOne from '../ctaOne/CtaOne'
-
-const News = () => {
-	return (
-		<>
-			<CtaOne />
-		</>
-	)
-}
-export default News;
